backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
numpy
opencv-python
ultralytics
scikit-learn
matplotlib
python-jose[cryptography]
google-auth